requirement.txt
absl-py==0.7.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
